+++ libs/abcvlib/src/main/java/jp/oist/abcvlib/core/inputs/microcontroller/BatteryDataListener.java
@@ -1,6 +1,6 @@
 package jp.oist.abcvlib.core.inputs.microcontroller;
 
 public interface BatteryDataListener {
-    void onBatteryVoltageUpdate(double voltage, double timestamp);
-    void onChargerVoltageUpdate(double voltage, double timestamp);
+    void onBatteryVoltageUpdate(double voltage, long timestamp);
+    void onChargerVoltageUpdate(double voltage, long timestamp);
 }
